[fix] / infinite scrolling error
